Fixed errors, restructured project.
Created a setup.py file and restructured the project to have
conventional structure.  This fixed import issues.

Decided to use a special xml tag attribute to set quote type.

Decided to remove reader macros so as to not pollute the client's
environment.

An issue with hyc cropped up reguarding imports from pyc; xmlhy.xmlhy
was not being imported. Resolved by using a hack, as noted in xmlhy.hy.